+++ ui-tests-protractor/e2e/features/connections/connections-kebab.feature
