+++ ui-tests-protractor/e2e/features/connections/delete-connections.feature
