+++ ui-tests-protractor/e2e/features/integrations/basic-filter.feature
