+++ ui-tests-protractor/e2e/features/integrations/delete-integrations.feature
