+++ ui-tests-protractor/e2e/features/integrations/detail-page-integrations.feature
