+++ ui-tests-protractor/e2e/features/screenshots/screenshots.feature
